Move manager origin logic to pallet_registrar
